[Refacto] refacto help_display function in multiple module
