+++ source/display/meson.build
@@ -1,5 +1,8 @@
 display_sources = files([
     'help_display.c',
+    'help_organizer.c',
+    'help_formatter.c',
+    'help_renderer.c',
     'usage_display.c',
     'version_display.c',
 ])
